refactor(api): move module command into own pkg
Refs: #57
