+++ api/components/bot_core/command/module/util.go
@@ -1,0 +1,93 @@
+/*
+ * JOJO Discord Bot - An advanced multi-purpose discord bot
+ * Copyright (C) 2022 Lazy Bytez (Elias Knodel, Pascal Zarrad)
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package module
+
+import (
+	"fmt"
+	"github.com/bwmarrin/discordgo"
+)
+
+// respondWithMissingComponent fills the passed discordgo.InteractionResponseData
+// with an embed field that indicates that the specified component could not be found.
+func respondWithMissingComponent(
+	s *discordgo.Session,
+	i *discordgo.InteractionCreate,
+	resp *discordgo.InteractionResponseData,
+	componentName interface{},
+) {
+	embeds := []*discordgo.MessageEmbedField{
+		{
+			Name:  ":x: Error",
+			Value: fmt.Sprintf("No module with name \"%v\" could be found!", componentName),
+		},
+	}
+
+	resp.Embeds[0].Fields = embeds
+
+	_ = s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
+		Type: discordgo.InteractionResponseChannelMessageWithSource,
+		Data: resp,
+	})
+}
+
+// respondWithAlreadyEnabled fills the passed discordgo.InteractionResponseData
+// with an embed field that indicates that the specified component is already enabled.
+func respondWithAlreadyEnabled(
+	s *discordgo.Session,
+	i *discordgo.InteractionCreate,
+	resp *discordgo.InteractionResponseData,
+	componentName interface{},
+) {
+	embeds := []*discordgo.MessageEmbedField{
+		{
+			Name:  ":x: Error",
+			Value: fmt.Sprintf("Module with name \"%v\" is already enabled!", componentName),
+		},
+	}
+
+	resp.Embeds[0].Fields = embeds
+
+	_ = s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
+		Type: discordgo.InteractionResponseChannelMessageWithSource,
+		Data: resp,
+	})
+}
+
+// respondWithAlreadyDisabled fills the passed discordgo.InteractionResponseData
+// with an embed field that indicates that the specified component is already disabled.
+func respondWithAlreadyDisabled(
+	s *discordgo.Session,
+	i *discordgo.InteractionCreate,
+	resp *discordgo.InteractionResponseData,
+	componentName interface{},
+) {
+	embeds := []*discordgo.MessageEmbedField{
+		{
+			Name:  ":x: Error",
+			Value: fmt.Sprintf("Module with name \"%v\" is already disabled!", componentName),
+		},
+	}
+
+	resp.Embeds[0].Fields = embeds
+
+	_ = s.InteractionRespond(i.Interaction, &discordgo.InteractionResponse{
+		Type: discordgo.InteractionResponseChannelMessageWithSource,
+		Data: resp,
+	})
+}
